+++ Assets/HUF/AdsAdMobMediation/SDK/GoogleMobileAds/Plugins/iOS/GoogleMobileAdsMediationTestSuite/GoogleMobileAdsMediationTestSuite+Internal.h
@@ -1,9 +1,0 @@
-// Copyright 2017 Google Inc. All Rights Reserved.
-
-#import <Foundation/Foundation.h>
-
-@interface GoogleMobileAdsMediationTestSuite (Internal)
-
-+ (void)setUserAgentSuffix:(NSString *)userAgentSuffix;
-
-@end