+++ Assets/InstantFramework/Scripts/Services/MoPub/TLAdsService.cs
@@ -3,11 +3,13 @@ using strange.extensions.promise.impl;
 using System.Collections.Generic;
 using HUF.Ads.Runtime.API;
 using HUF.Ads.Runtime.Implementation;
-using HUF.AdsAdMobMediation.Runtime.API;
+//using HUF.AdsAdMobMediation.Runtime.API;
 using HUFEXT.AdsManager.Runtime.API;
 using HUFEXT.AdsManager.Runtime.AdManagers;
 using TurboLabz.TLUtils;
-using HUF.AdsAdMobMediation.Runtime.Implementation;
+//using HUF.AdsAdMobMediation.Runtime.Implementation;
+using HUF.AdsIronSourceMediation.Runtime.Implementation;
+using HUF.AdsIronSourceMediation.Runtime.API;
 
 namespace TurboLabz.InstantFramework
 {
@@ -42,7 +44,7 @@ namespace TurboLabz.InstantFramework
             HAds.Banner.OnFailed += OnBannerFailed;
             HAds.Interstitial.OnClicked += OnInterstitialClicked;
             HAds.Rewarded.OnClicked += OnRewardedClicked;
-            HAdsAdMobMediation.OnPaidEvent += HandlePaidEvent;
+            //AdsIronSourceMediation.OnPaidEvent += HandlePaidEvent;
             HAdsManager.SetNewBannerPosition(BannerPosition.TopCenter);
         }
 
@@ -256,7 +258,7 @@ namespace TurboLabz.InstantFramework
 
         public void ShowTestSuite()
         {
-            HAdsAdMobMediation.ShowTestSuite();
+            HAdsManager.ShowAd(PLACEMENT_ID_INTERSTITIAL, null);
         }
 
         public void OnAppEvent(AppEvent evt)
@@ -270,12 +272,6 @@ namespace TurboLabz.InstantFramework
             //    analyticsService.Event(AnalyticsEventId.ad_player_shutdown, playerModel.adContext);
             //}
         }
-
-        private void HandlePaidEvent(PaidEventData data)
-        {
-            hAnalyticsService.LogAdImpressionEvent(data);
-        }
-
 
         private void OnBannerFailed(IBannerCallbackData data)
         {